+++ hooks/queries/index.tsx
@@ -2,12 +2,10 @@ import { useGetAttributes } from "./getAttributes";
 import { useGetVenues } from "./getVenues";
 import { useGetVenuesByAttributes } from "./getVenuesByAttributes";
 import { useGetVenuesByName } from "./getVenuesByName";
-import { useGetMapBoxLocations } from "./getMapBoxLocations";
 
 export {
   useGetAttributes,
   useGetVenues,
   useGetVenuesByName,
   useGetVenuesByAttributes,
-  useGetMapBoxLocations,
 };
